+++ emulator.hpp
@@ -11,7 +11,7 @@ uint16_t byteAtIndex(std::uint16_t, int, int = 0);
 
 struct CallInfo{
     std::vector<uint8_t>::iterator returnAddress;
-    std::vector<uint16_t> registers;
+    std::vector<uint8_t> registers;
 };
 
 class Emulator
@@ -25,7 +25,7 @@ public:
 
     bool isEnd();
 
-    std::vector<std::uint16_t>& getRegisters();
+    std::vector<std::uint8_t>& getRegisters();
     std::vector<std::uint8_t>& getMemory();
     std::vector<std::uint8_t>::iterator getProgramCounter();
     std::array<bool, 2048>& getDisplayBuffer();
@@ -35,7 +35,7 @@ public:
 private:
     std::vector<std::uint8_t> m_memory;
     std::stack<CallInfo> m_callStack;
-    std::vector<std::uint16_t> m_registers;
+    std::vector<std::uint8_t> m_registers;
     std::uint16_t m_registerI;
     std::array<bool, 2048> m_displayBuffer;
     std::vector<std::uint8_t>::iterator m_progCounter;
